Reorganize
Load images dynamically instead of loading them all in a carousel.
Split code into submodules.
