+++ assets/texts/paragraph1.txt
@@ -8,6 +8,6 @@ Accommodating all of these goals and priorities within a single app would be ver
 
    • Aims to strengthen movement, discipline, and muscle memory rather than teach specific sequences.
 
-   • Accommodates newcomers more than experienced practitioners (although experienced practitioners could find this app useful for solo practice).
+   • Targets newcomers more than experienced practitioners (although experienced practitioners could find this app useful for solo practice).
 
-While every historical text has its own unique approach and emphasis, nearly all of those texts share certain commonalities. For example, the following images (scroll to the right to see more) range in time period from the 1400s to modern day, come from Italian, German, British, Polish, Japanese, and American instructional texts, and apply to longsword, dussak, sabre, and other weapons:
+While every historical text has its own unique approach and emphasis, nearly all of those texts share certain commonalities. For example, the following images range in time period from the 1400s to modern day, come from Italian, German, British, Polish, Japanese, and American instructional texts, and apply to longsword, dussak, sabre, and other weapons:
